Move main gutenberg entry definition and webpack copy plugin out of build-config.
